# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hexagonal Go service template with MongoDB persistence adapter.",
        "title": "Golang Service Template API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/gambitier/golang-service-template"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/items/api/v1",
    "paths": {
        "/items": {
            "get": {
                "description": "List items with optional pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "List items",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create item",
                "parameters": [
                    {
                        "description": "Item payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presentation_http_handlers_item.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "description": "Fetch a single item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get item by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Update item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presentation_http_handlers_item.UpdateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an item by ID",
                "tags": [
                    "items"
                ],
                "summary": "Delete item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_gambitier_golang-service-template_internal_presentation_http_response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "fields": {},
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "success": {
                    "type": "integer"
                }
            }
        },
        "internal_presentation_http_handlers_item.CreateItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A ruled notebook"
                },
                "name": {
                    "type": "string",
                    "example": "Notebook"
                }
            }
        },
        "internal_presentation_http_handlers_item.UpdateItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "name": {
                    "type": "string",
                    "example": "Notebook"
                }
            }
        }
    }
}